Change folder name, delete clicker tutorial, fix script error at Click.cs
