--- static/unity/Assets/Clicker Game/Scripts/Menu.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class Menu : MonoBehaviour {
-
-
-	public void StartGame () {
-		Application.LoadLevel (1);
-	}
-	
-
-	public void Quit () {
-		Application.Quit();
-	}
-}
