--- static/unity/Assets/Clicker Game/Scripts/Pause.cs
@@ -1,12 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class Pause : MonoBehaviour {
-
-		
-	void Update () {
-	if (Input.GetKey (KeyCode.Escape)) {
-			Application.LoadLevel(0);
-		}
-	}
-}
