--- static/unity/Assets/Clicker Game/_Tutorials (Video and .TXT)/Tutorial.txt (How to use the kit).txt
@@ -1,34 +1,0 @@
-1. Scripts--
-    -1: Click.cs : If player click on the cookie he will get + 1 cookie. In this script store the
-    Texts, player's cookie quantity. In this script we write the current cookie quantity to
-    the screen(UI text).(So the CPC)
-    -2: CurrencyManager.cs : This script convert the cookie. So if player get 1000 cookie is
-    converted to 1K if player get 1000000 cookie its converted to 1M (at 1500 = 1.5K its a float).
-    Biggest value is 1000000000000000000 = Qui (Its the maximum cookie).
-    -3: GoldPerSec.cs : In this script we store how many cookies get the player / sec (So the CPS).
-    -4: ItemManager.cs : This is manage the CPS items. We store here the item's: cost, tickValue(sec/c),
-    -5: UpgradeManager.cs : This is manage the CPC items.
-    -6: Menu.cs : A simple menu script.
-    -7: Pause.cs : If the player press ESC menu is show up.
-    
-    
-2. How to add items.
-	-1: First you need to choose item type CPS(cookie per sec) or CPC(cookie per click) item.
-		-2/a: If you choose CPC you will need the UpgradeManager script.
-		-2/b: If you choose CPS you will need the ItemManager script.
-	-3: Okay if you choosen click on Canvas (see childrens).
-	-4: If you choose CPC click on itemMenu_GPC if you choose CPS itemMenu_CPS and see the childrens
-	-5: If you click on the arrow you see Scroll its the scroll menu to add infinite items. Click on arrow again 
-	to see scroll menu childrens. 
-	-6: Now duplicate the last item and rename it. (if last item name cps_2 rename to cps_3)-
-	-7: Now make your item. You see the inspector the Item Manager script now you need configure this.
-		-8/a: If you choose CPC: Give a cost for the item (integer) / count live it 0 / Click power (cookie per click)
-		/ Item Name  give any name for your item.	
-		-8/b: If you choose CPS: Give a cost for the item (integer) / Give a tick value  to item (How many cookie per sec)
-		/count leave it 0 / Item Name  give any name for your item.
-	-9: Now your item is ready.
-	
-	
-If you have any question my email: monofoxdevs@gmail.hu
-My skype: monofox.tomi 
-I will answer you in 1-2 days. :D
