+++ static/unity/Assets/ClickerGame/Scripts/ItemManager.cs
@@ -1,0 +1,57 @@
+﻿using UnityEngine;
+using System.Collections;
+using UnityEngine.UI;
+public class ItemManager : MonoBehaviour {
+
+	public UnityEngine.UI.Text itemInfo;
+	public Click click;
+	public float cost;
+	public int tickValue;
+	public int count;
+	public string itemName;
+	private float baseCost;
+	public Color standard;
+	public Color affordable;
+	private Slider _slider;
+
+
+
+	void Start(){
+		baseCost = cost; 
+		_slider = GetComponentInChildren<Slider> ();
+	}
+
+
+
+	void Update(){
+
+
+
+
+		itemInfo.text = itemName + "\nCost: "+cost+"\nGold: "+tickValue+"/s";
+
+		/*if (click.gold >= cost) {
+			GetComponent<Image> ().color = affordable;
+		} else {
+			GetComponent<Image> ().color = standard;
+			
+		} 
+		*/
+
+		_slider.value = click.gold / cost * 100;
+
+}
+
+
+	public void PurchasedItem(){
+		if (click.gold >= cost) {
+			click.gold -= cost;
+			count += 1;
+			cost = Mathf.Round( baseCost * Mathf.Pow (click.costmultiplier, count));
+
+		}
+	
+	
+	}
+
+}
