+++ static/unity/Assets/ClickerGame/Scripts/Menu.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class Menu : MonoBehaviour {
+
+
+	public void StartGame () {
+		Application.LoadLevel (1);
+	}
+	
+
+	public void Quit () {
+		Application.Quit();
+	}
+}
