+++ static/unity/Assets/ClickerGame/Scripts/Pause.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class Pause : MonoBehaviour {
+
+		
+	void Update () {
+	if (Input.GetKey (KeyCode.Escape)) {
+			Application.LoadLevel(0);
+		}
+	}
+}
